+++ src/components/CurrencyInput.tsx
@@ -1,7 +1,7 @@
-import FormField from '@components/FormField';
+import FormField from '@internal/FormField';
+import type { IFieldProps } from '@internal/types';
+import { describeBy } from '@internal/utils';
 import { useSignal } from '@preact/signals';
-import type { IFieldProps } from '@scripts/types';
-import { describeBy } from '@scripts/utils';
 import styles from '@styles/CurrencyInput.module.css';
 import type { FunctionComponent } from 'preact';
 import { useId } from 'preact/hooks';