+++ src/components/Input.tsx
@@ -1,6 +1,6 @@
-import FormField from '@components/FormField';
-import type { IFieldProps } from '@scripts/types';
-import { describeBy } from '@scripts/utils';
+import FormField from '@internal/FormField';
+import type { IFieldProps } from '@internal/types';
+import { describeBy } from '@internal/utils';
 import styles from '@styles/Input.module.css';
 import type { FunctionComponent } from 'preact';
 import { useId } from 'preact/hooks';