+++ src/components/NotFound.tsx
@@ -1,7 +1,7 @@
 import Link from '@components/Link';
-import type { IBaseProps } from '@scripts/types';
-import { cx } from '@scripts/utils';
+import type { IBaseProps } from '@internal/types';
 import styles from '@styles/NotFound.module.css';
+import { cx } from '@utils';
 import type { FunctionComponent } from 'preact';
 
 /** Props for {@link NotFound}. */