+++ src/components/Textarea.tsx
@@ -1,8 +1,8 @@
-import FormField from '@components/FormField';
 import VisuallyHidden from '@components/VisuallyHidden';
+import FormField from '@internal/FormField';
+import type { IFieldProps } from '@internal/types';
+import { describeBy } from '@internal/utils';
 import { useSignal } from '@preact/signals';
-import type { IFieldProps } from '@scripts/types';
-import { describeBy } from '@scripts/utils';
 import styles from '@styles/Textarea.module.css';
 import type { FunctionComponent } from 'preact';
 import { useId } from 'preact/hooks';